backend/requirements.txt
Flask
tensorflow
polygon-api-client
pandas
numpy
scikit-learn
